# controller/dashboard.go
package controller

import (
	"context"
	"encoding/json"
	"net/http"
	"strings"
	"time"

	"github.com/golang-jwt/jwt/v5"
	"github.com/kawai-org/kawai-backend/config"
	"github.com/kawai-org/kawai-backend/model"
	"go.mongodb.org/mongo-driver/bson"
	"go.mongodb.org/mongo-driver/bson/primitive"
	"go.mongodb.org/mongo-driver/mongo/options"
)

// Helper: Ambil Claims dari Context
func getUserClaims(r *http.Request) jwt.MapClaims {
	claims, _ := r.Context().Value("user").(jwt.MapClaims)
	return claims
}

// Helper: Ambil ID dari URL
func getIDFromURL(r *http.Request, prefix string) (primitive.ObjectID, error) {
	idStr := strings.TrimPrefix(r.URL.Path, prefix)
	return primitive.ObjectIDFromHex(idStr)
}

// ==========================================
// BAGIAN 1: CRUD USER (NOTES & REMINDERS)
// ==========================================

// --- NOTES ---

// GetMyNotes godoc
// @Summary Ambil Catatan User
// @Description Mendapatkan daftar catatan milik pengguna yang sedang login.
// @Tags Dashboard
// @Produce json
// @Param search query string false "Keyword Pencarian"
// @Success 200 {object} map[string]interface{}
// @Security BearerAuth
// @Router /api/dashboard/notes [get]
func GetMyNotes(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	claims := getUserClaims(r)
	userPhone, _ := claims["user_phone"].(string)

	// Ambil parameter ?search=... dari URL
	keyword := r.URL.Query().Get("search")

	// Filter Dasar: Harus milik user ini
	filter := bson.M{"user_phone": userPhone}

	// Jika ada keyword search, tambahkan filter regex (case insensitive)
	if keyword != "" {
		filter["content"] = bson.M{
			"$regex":   keyword,
			"$options": "i", // "i" artinya tidak peduli huruf besar/kecil
		}
	}

	opts := options.Find().SetSort(bson.M{"created_at": -1})

	cursor, err := config.Mongoconn.Collection("notes").Find(context.TODO(), filter, opts)
	if err != nil {
		json.NewEncoder(w).Encode(map[string]interface{}{"status": "error", "msg": "DB Error"})
		return
	}

	notes := []model.Note{}
	if cursor != nil {
		cursor.All(context.TODO(), &notes)
	}

	json.NewEncoder(w).Encode(map[string]interface{}{
		"status": "success",
		"data":   notes,
	})
}

// GetNoteDetailWithTags godoc
// @Summary Detail Catatan
// @Description Melihat detail catatan beserta hashtagnya
// @Tags Dashboard
// @Produce json
// @Param id path string true "Note ID"
// @Success 200 {object} map[string]interface{}
// @Security BearerAuth
// @Router /api/dashboard/notes/{id}/detail [get]
func GetNoteDetailWithTags(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	
    // Ambil ID note dari URL
    noteID, err := getIDFromURL(r, "/api/notes/detail/")
	if err != nil {
		json.NewEncoder(w).Encode(map[string]string{"status": "error", "msg": "ID URL Salah"})
		return
	}

	// Pipeline Aggregation (Join notes + tags)
	// KITA GANTI bson.D JADI bson.M AGAR TIDAK KUNING
	pipeline := bson.A{
		// 1. Filter Note berdasarkan ID
		bson.M{"$match": bson.M{"_id": noteID}},

		// 2. JOIN ke tabel tags
		bson.M{"$lookup": bson.M{
			"from":         "tags",      
			"localField":   "_id",       // Field di collection 'notes'
			"foreignField": "note_id",   // Field di collection 'tags'
			"as":           "related_tags", // Nama field baru untuk hasil join
		}},
	}

	cursor, err := config.Mongoconn.Collection("notes").Aggregate(context.TODO(), pipeline)
    if err != nil {
         json.NewEncoder(w).Encode(map[string]string{"status": "error", "msg": "Gagal mengambil data"})
		 return
    }
    
    // Decode hasilnya
    var results []bson.M
    if cursor != nil {
        cursor.All(context.TODO(), &results)
    }
    
	// Cek apakah data ditemukan
	if len(results) == 0 {
		json.NewEncoder(w).Encode(map[string]string{"status": "error", "msg": "Catatan tidak ditemukan"})
		return
	}

    // Kembalikan item pertama (karena search by ID pasti cuma 1)
    json.NewEncoder(w).Encode(map[string]interface{}{
		"status": "success",
		"data":   results[0],
	})
}

// UpdateNote godoc
// @Summary Update Catatan
// @Description Mengubah isi catatan
// @Tags Dashboard
// @Accept json
// @Produce json
// @Param id path string true "Note ID"
// @Param input body map[string]string true "Content baru"
// @Success 200 {object} map[string]string
// @Security BearerAuth
// @Router /api/notes/{id} [put]
func UpdateNote(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	claims := getUserClaims(r)
	userPhone, _ := claims["user_phone"].(string)

	noteID, err := getIDFromURL(r, "/api/notes/")
	if err != nil {
		json.NewEncoder(w).Encode(map[string]string{"status": "error", "msg": "ID URL Salah"})
		return
	}

	var input struct { Content string `json:"content"` }
	json.NewDecoder(r.Body).Decode(&input)

	filter := bson.M{"_id": noteID, "user_phone": userPhone}
	update := bson.M{"$set": bson.M{"content": input.Content}}

	res, _ := config.Mongoconn.Collection("notes").UpdateOne(context.TODO(), filter, update)
	if res.MatchedCount == 0 {
		json.NewEncoder(w).Encode(map[string]string{"status": "error", "msg": "Gagal update"})
		return
	}
	json.NewEncoder(w).Encode(map[string]string{"status": "success", "msg": "Updated"})
}

// DeleteNote godoc
// @Summary Hapus Catatan
// @Description Menghapus catatan berdasarkan ID
// @Tags Dashboard
// @Produce json
// @Param id path string true "Note ID"
// @Success 200 {object} map[string]string
// @Security BearerAuth
// @Router /api/notes/{id} [delete]
func DeleteNote(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	claims := getUserClaims(r)
	userPhone, _ := claims["user_phone"].(string)

	noteID, err := getIDFromURL(r, "/api/notes/")
	if err != nil {
		json.NewEncoder(w).Encode(map[string]string{"status": "error", "msg": "ID URL Salah"})
		return
	}

	filter := bson.M{"_id": noteID, "user_phone": userPhone}
	res, _ := config.Mongoconn.Collection("notes").DeleteOne(context.TODO(), filter)
	
	if res.DeletedCount == 0 {
		json.NewEncoder(w).Encode(map[string]string{"status": "error", "msg": "Gagal hapus"})
		return
	}
	json.NewEncoder(w).Encode(map[string]string{"status": "success", "msg": "Deleted"})
}

// --- REMINDERS ---

// --- REMINDERS ---

// GetReminders godoc
// @Summary Ambil Agenda Pengingat
// @Description Mengambil daftar pengingat yang akan datang.
// @Tags Dashboard
// @Produce json
// @Param search query string false "Keyword Pencarian"
// @Success 200 {object} map[string]interface{}
// @Security BearerAuth
// @Router /api/dashboard/reminders [get]
func GetReminders(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	claims := getUserClaims(r)
	userPhone, _ := claims["user_phone"].(string)

	// 1. Ambil Keyword Search (Tambahan Baru)
	keyword := r.URL.Query().Get("search")

	filter := bson.M{"user_phone": userPhone}

	// 2. Logika Search Regex (Cari berdasarkan Judul)
	if keyword != "" {
		filter["title"] = bson.M{
			"$regex":   keyword,
			"$options": "i", // Case insensitive (Huruf besar/kecil dianggap sama)
		}
	}

	// Sort berdasarkan waktu reminder terdekat
	opts := options.Find().SetSort(bson.M{"scheduled_time": 1})

	cursor, err := config.Mongoconn.Collection("reminders").Find(context.TODO(), filter, opts)
	if err != nil {
		json.NewEncoder(w).Encode(map[string]string{"status": "error", "msg": "DB Error"})
		return
	}

	reminders := []model.Reminder{}
	if cursor != nil {
		cursor.All(context.TODO(), &reminders)
	}

	json.NewEncoder(w).Encode(map[string]interface{}{
		"status": "success", 
		"data": reminders,
	})
}

// UpdateReminder godoc
// @Summary Update Reminder
// @Description Mengubah Judul atau Waktu Pengingat
// @Tags Dashboard
// @Accept json
// @Produce json
// @Param id path string true "Reminder ID"
// @Param input body map[string]interface{} true "Title / ScheduledTime"
// @Success 200 {object} map[string]string
// @Security BearerAuth
// @Router /api/reminders/{id} [put]
func UpdateReminder(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	claims := getUserClaims(r)
	userPhone, _ := claims["user_phone"].(string)

	id, err := getIDFromURL(r, "/api/reminders/")
	if err != nil {
		json.NewEncoder(w).Encode(map[string]string{"status": "error", "msg": "ID Salah"})
		return
	}

	var input struct {
		Title string    `json:"title"`
		Time  time.Time `json:"scheduled_time"`
	}
	json.NewDecoder(r.Body).Decode(&input)

	updateFields := bson.M{}
	if input.Title != "" { updateFields["title"] = input.Title }
	if !input.Time.IsZero() { updateFields["scheduled_time"] = input.Time }

	filter := bson.M{"_id": id, "user_phone": userPhone}
	res, _ := config.Mongoconn.Collection("reminders").UpdateOne(context.TODO(), filter, bson.M{"$set": updateFields})

	if res.MatchedCount == 0 {
		json.NewEncoder(w).Encode(map[string]string{"status": "error", "msg": "Gagal update"})
	} else {
		json.NewEncoder(w).Encode(map[string]string{"status": "success", "msg": "Updated"})
	}
}

// [BARU] DELETE REMINDER
// DeleteReminder godoc
// @Summary Hapus Reminder
// @Description Menghapus pengingat berdasarkan ID
// @Tags Dashboard
// @Produce json
// @Param id path string true "Reminder ID"
// @Success 200 {object} map[string]string
// @Security BearerAuth
// @Router /api/reminders/{id} [delete]
func DeleteReminder(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	claims := getUserClaims(r)
	userPhone, _ := claims["user_phone"].(string)

	id, err := getIDFromURL(r, "/api/reminders/")
	if err != nil {
		json.NewEncoder(w).Encode(map[string]string{"status": "error", "msg": "ID Salah"})
		return
	}

	filter := bson.M{"_id": id, "user_phone": userPhone}
	res, _ := config.Mongoconn.Collection("reminders").DeleteOne(context.TODO(), filter)

	if res.DeletedCount == 0 {
		json.NewEncoder(w).Encode(map[string]string{"status": "error", "msg": "Gagal hapus"})
	} else {
		json.NewEncoder(w).Encode(map[string]string{"status": "success", "msg": "Deleted"})
	}
}

// ==========================================
// BAGIAN 2: ADMIN SIMPLE (Tukang Intip & Satpam)
// ==========================================

// Helper Cek Admin
func isAdmin(r *http.Request) bool {
	claims := getUserClaims(r)
	role, ok := claims["role"].(string)
	return ok && role == "admin" 
}

// 1. LIHAT SEMUA USER (Untuk Audit/Monitoring)
// GetAllUsers godoc
// @Summary [ADMIN] Lihat Users
// @Description Mengambil semua data user yang terdaftar
// @Tags Admin
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Security BearerAuth
// @Router /api/admin/users [get]
func GetAllUsers(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	if !isAdmin(r) {
		w.WriteHeader(http.StatusForbidden)
		json.NewEncoder(w).Encode(map[string]string{"status": "error", "msg": "Khusus Admin"})
		return
	}

	cursor, _ := config.Mongoconn.Collection("users").Find(context.TODO(), bson.M{})
	users := []model.User{}
	if cursor != nil { cursor.All(context.TODO(), &users) }

	json.NewEncoder(w).Encode(map[string]interface{}{"status": "success", "total": len(users), "data": users})
}

// 2. STATISTIK SIMPLE (Total User, Total Notes)
// GetSystemStats godoc
// @Summary [ADMIN] Lihat Statistik
// @Description Mengambil statistik sederhana sistem
// @Tags Admin
// @Produce json
// @Success 200 {object} map[string]interface{}
// @Security BearerAuth
// @Router /api/admin/stats [get]
func GetSystemStats(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	if !isAdmin(r) {
		w.WriteHeader(http.StatusForbidden)
		return
	}

	userCount, _ := config.Mongoconn.Collection("users").CountDocuments(context.TODO(), bson.M{})
	noteCount, _ := config.Mongoconn.Collection("notes").CountDocuments(context.TODO(), bson.M{})
	reminderCount, _ := config.Mongoconn.Collection("reminders").CountDocuments(context.TODO(), bson.M{})

	json.NewEncoder(w).Encode(map[string]interface{}{
		"status": "success",
		"stats": map[string]int64{
			"total_users":     userCount,
			"total_notes":     noteCount,
			"total_reminders": reminderCount,
		},
	})
}

// 3. BAN USER (Blokir User Iseng)
// BanUser godoc
// @Summary [ADMIN] Ban User
// @Description Memblokir atau membuka blokir user
// @Tags Admin
// @Accept json
// @Produce json
// @Param input body map[string]string true "Target Phone & Action (ban/unban)"
// @Success 200 {object} map[string]string
// @Security BearerAuth
// @Router /api/admin/ban [post]
func BanUser(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	if !isAdmin(r) {
		w.WriteHeader(http.StatusForbidden)
		return
	}

	var input struct {
		Phone  string `json:"phone_number"`
		Action string `json:"action"` // "ban" atau "unban"
	}
	json.NewDecoder(r.Body).Decode(&input)

	status := "active"
	if input.Action == "ban" {
		status = "banned"
	}

	filter := bson.M{"phone_number": input.Phone}
	update := bson.M{"$set": bson.M{"status": status}}
	
	res, _ := config.Mongoconn.Collection("users").UpdateOne(context.TODO(), filter, update)
	
	if res.MatchedCount == 0 {
		json.NewEncoder(w).Encode(map[string]string{"status": "error", "msg": "User tidak ditemukan"})
	} else {
		json.NewEncoder(w).Encode(map[string]string{"status": "success", "msg": "Status user diupdate: " + status})
	}
}