# docs/swagger.yaml
basePath: /
definitions:
  controller.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
host: kawai-be.vercel.app
info:
  contact:
    name: Dwi Puspa Firdaus
  description: Dokumentasi API Backend untuk Kawai Assistant (Bot WhatsApp & Dashboard).
  title: Kawai Assistant API
  version: 1.0.0
paths:
...
  /api/admin/ban:
    post:
      consumes:
      - application/json
      description: Memblokir atau membuka blokir user
      parameters:
      - description: Target Phone & Action (ban/unban)
        in: body
        name: input
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: '[ADMIN] Ban User'
      tags:
      - Admin
  /api/admin/login:
    post:
      consumes:
      - application/json
      description: Login untuk admin menggunakan nomor HP dan password
      parameters:
      - description: Kredensial Login
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controller.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: Token JWT dan Role
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid Body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Password Salah / Bukan Admin
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login Admin
      tags:
      - Auth
  /api/admin/stats:
    get:
      description: Mengambil statistik sederhana sistem
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: '[ADMIN] Lihat Statistik'
      tags:
      - Admin
  /api/admin/users:
    get:
      description: Mengambil semua data user yang terdaftar
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: '[ADMIN] Lihat Users'
      tags:
      - Admin
...
  /api/dashboard/notes:
    get:
      description: Mendapatkan daftar catatan milik pengguna yang sedang login.
      parameters:
      - description: Keyword Pencarian
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Ambil Catatan User
      tags:
      - Dashboard
  /api/dashboard/notes/{id}/detail:
    get:
      description: Melihat detail catatan beserta hashtagnya
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Detail Catatan
      tags:
      - Dashboard
  /api/dashboard/reminders:
    get:
      description: Mengambil daftar pengingat yang akan datang.
      parameters:
      - description: Keyword Pencarian
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Ambil Agenda Pengingat
      tags:
      - Dashboard
  /api/notes/{id}:
    delete:
      description: Menghapus catatan berdasarkan ID
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Hapus Catatan
      tags:
      - Dashboard
    put:
      consumes:
      - application/json
      description: Mengubah isi catatan
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: string
      - description: Content baru
        in: body
        name: input
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Catatan
      tags:
      - Dashboard
...
  /api/reminders/{id}:
    delete:
      description: Menghapus pengingat berdasarkan ID
      parameters:
      - description: Reminder ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Hapus Reminder
      tags:
      - Dashboard
    put:
      consumes:
      - application/json
      description: Mengubah Judul atau Waktu Pengingat
      parameters:
      - description: Reminder ID
        in: path
        name: id
        required: true
        type: string
      - description: Title / ScheduledTime
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Reminder
      tags:
      - Dashboard
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"